Fix wrong-file bug in fs persistence plugin
The savePlayer function was reading from the wrong file (tasks) and
then merging the new player. The test didn't catch it because the
'emptyStates' object was being mutated, such that as long as the
object stayed constructed (which is true during the test, but wouldn't
be in real life) data could appear to 'leak' between the tasks and
players files.