Support payment-type currency changes
My hope is that this way of handling currency exchanges will scale
well when dealing with race conditions; a transaction can get added to
the ledger, but if at that point in the ledger it results in negative
currency for a cost being paid, it's simply ignored when projecting
the current state.

This suggests a kind of eventual consistency; someone trying to
purchase an asset, for example, might submit the transaction, then
poll sometime later to see if the asset is showing up in the player's
current state. If not, the transaction must have been rejected.