Create rrule plugin, implement findNextOccurrence
I've decided for now to leave construction of the complete
iCal object (including DTSTART and RRULE lines) up to
the caller of createRecurringTask.

This makes testing a little tricky, but it's more flexible than
other approaches (like storing a createdAt time on
the recurring task and reusing that) would be.

If it turns out to be more trickery than it's worth, though,
using a createdAt time on the recurring task might be a better play.

The timezone handling between the application and the rrule library
feels...precarious. Seems to work like you'd expect in testing.
Something to watch out for, though.