Don't let findNextOccurrence return the baseTime
findNextOccurrence always needs to return a time _after_ the supplied
base time.

This way, I can pass the createdAt time of the latest instance of a
recurring task into findNextOccurrence to find out the next time the
task should spawn. If findNextOccurrence was allowed to return the
base time when the base time fell exactly on an occurrence, then
passing in the createdAt time of a recurring task instance would
always return that time (at least in cases where the task was spawned
exactly on schedule).